+++ app/src/main/java/com/example/pboguet/macaveonline/Utils/Utilisateur.java
@@ -4,14 +4,14 @@ package com.example.pboguet.macaveonline.Utils;
  * Created by pboguet on 16/04/15.
  */
 public class Utilisateur {
-    public long userId;
+    public int userId;
     public String login;
     public String password;
     public String nom;
     public String prenom;
     public String mail;
 
-    public Utilisateur(long userId, String username, String password, String nom, String prenom, String mail){
+    public Utilisateur(int userId, String username, String password, String nom, String prenom, String mail){
         this.userId=userId;
         this.login=username;
         this.password=password;
@@ -20,11 +20,11 @@ public class Utilisateur {
         this.mail = mail;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
